+++ src/main/gen/arma/config/psi/MacroAtom.java
@@ -13,4 +13,7 @@ public interface MacroAtom extends PsiElement {
   @Nullable
   MacroInnerToken getMacroInnerToken();
 
+  @Nullable
+  SingleQuoteBlock getSingleQuoteBlock();
+
 }