+++ src/main/resources/example.hpp
@@ -13,49 +13,48 @@ comment
 multiline
 */
 
-//todo works?
-//fixme works as well?
-
-
-
-class 30Rnd_556x45_Stanag;
+class SingleLineClass { MACRO(include); };
+class 01ClassStartingWithNumber;
 class NoBodyClass;
-class ErrorsShouldBeShowingHere {
+class InspectionTestingClass {
     TEST(var,var);
     arrayWrong = {};
     arrayWrond2[] = test;
     duplicatedParam = 1;
     duplicatedParam = 1;
-}
-class TestClass {
+};
+class NormalClass {
     condition = Q(count TAG(targets) > 0);
     condition2 = Q(!(_target in TAG(targets)));
 
-	indentifier = anotherIdentifer;
-	array[] = {}; //inline comment
-	arrayMultiline[] = {
-	    "test",
+    indentifier = anotherIdentifer;
+    array[] = {}; //inline comment
+    arrayMultiline[] = {
+        "test",
         "test"
-        ,"test"
-	}; //inline comment
-	string = "hallo from the other siiide";
+        , "test"
+    }; //inline comment
+    string = "hallo from the other siiide";
+    string = "Here is a ""substring""";
+    string = 'Here is a "substring"';
+    string = "Here is a 'substring'";
     TEST(var,var);
-	math = 1 + 1 + (42 - 2);
-	float = 2.20;
-	number = 12;
+    math = 1 + 1 + (42 - 2);
+    float = 2.20;
+    number = 12;
     preprocessorThing = Q(path\to\file.paa);
     complexPreprocessor = Q(_target setObjectTextureGlobal [0, 'PATH(textures\ipsc_vertical.paa)'];);
     class NestedClass : extendClass {
-		/*Another block comment*/
-		array[] = {var, 1, 1, "", 42, 43};
-		math = 1 + 69 * 1;
-	};
-	class 30Rnd_556x45_Stanag;
-	class test : 30Rnd_556x45_Stanag {};
+        /*Another block comment*/
+        array[] = {var, 1, 1, "", 42, 43};
+        math = 1 + 69 * 1;
+    };
+    class 30Rnd_556x45_Stanag;
+    class test : 30Rnd_556x45_Stanag {};
     class TAG(preprocessorClass) : TAG(extensionAlso) {};
     thingy = 0;
-	delete DeleteIsAKeywordTooClass;
-	PREPROCESSOR_USED_FOR_PARAMS(something);
-	JUST_A_PREPROCESSOR_THING;
-	class NoBodyClass;
+    delete DeleteIsAKeywordTooClass;
+    PREPROCESSOR_USED_FOR_PARAMS(something);
+    JUST_A_PREPROCESSOR_THING;
+    class NoBodyClass;
 };